Implement metric/imperial units, add support for Tang EV, add in-app changelog viewer, implement more trip info inside trip recording card, add SoC into battery chart, fix enginePower to Int

+++ app/src/androidTest/java/com/byd/tripstats/data/test/TelemetryFactory.kt
@@ -10,7 +10,7 @@ import java.time.Instant
 fun telemetry(
     gear          : String = "P",
     speed         : Double = 0.0,
-    enginePower   : Double = 0.0,
+    enginePower   : Int    = 0,
     chargingPower : Double = 0.0,
     soc           : Double = 80.0,
     carOn         : Int    = 1,
